% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{calculate_delay_ratio}
\alias{calculate_delay_ratio}
\title{Calculates the change in the delay log-likelihood}
\usage{
calculate_delay_ratio(
  proposed_dat,
  current_dat,
  i,
  group_idx,
  delay_idx,
  theta,
  obs_dat,
  index_dates
)
}
\arguments{
\item{proposed_dat}{Proposed augmented data (e.g. `proposed_aug_dat_step3`)}

\item{current_dat}{Current augmented data (state before the new proposal)}

\item{i}{Index for individual}

\item{group_idx}{Index for group}

\item{delay_idx}{Vector of delay indices affected by the date changes}

\item{theta}{List of model parameters, including `mu` and `CV`}

\item{obs_dat}{List of observed data}

\item{index_dates}{List defining the delays}
}
\value{
Numeric value for the log-ratio of the delay likelihoods
}
\description{
Calculate the change in the delay component of the log-likelihood
between a proposed and current state.
}
