% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{calculate_error_ratio}
\alias{calculate_error_ratio}
\title{Calculates the change in the error log-likelihood}
\usage{
calculate_error_ratio(
  proposed_dat,
  current_dat,
  i,
  group_idx,
  all_changed_dates,
  theta
)
}
\arguments{
\item{proposed_dat}{Proposed augmented data (e.g. `proposed_aug_dat_step3`)}

\item{current_dat}{Current augmented data (state before the new proposal)}

\item{i}{Index for individual}

\item{group_idx}{Index for group}

\item{all_changed_dates}{Vector of all date indices that were modified}

\item{theta}{List of model parameters, including `zeta`.}
}
\value{
Numeric value for the log-ratio of the error likelihoods
}
\description{
Calculate the change in the error component of the log-likelihood
between a proposed and current state.
}
