% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{calculate_obs_ratio}
\alias{calculate_obs_ratio}
\title{Calculate the change in observation log-likelihood}
\usage{
calculate_obs_ratio(
  proposed_dat,
  current_dat,
  i,
  group_idx,
  all_changed_dates,
  obs_dat,
  range_dates
)
}
\arguments{
\item{proposed_dat}{Proposed augmented data (e.g. `proposed_aug_dat_step3`)}

\item{current_dat}{Current augmented data (state before the new proposal)}

\item{i}{Index for individual}

\item{group_idx}{Index for group}

\item{all_changed_dates}{Vector of all date indices that were modified}

\item{obs_dat}{List of observed data}

\item{range_dates}{Vector of overall date range}
}
\value{
Numeric value of the log-ratio of observation likelihoods
}
\description{
Calculate the change in the observation component of the log-likelihood
between a proposed and current state.
}
