% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{check_convergence}
\alias{check_convergence}
\title{Check convergence of MCMC chain}
\usage{
check_convergence(x, threshold = 1.1)
}
\arguments{
\item{x}{MCMC chain, this is \code{theta_chain} from \code{RunMCMC}}

\item{threshold}{numeric, set to 1.1. This can be configured
but it is probably best not to change it. Gelman-Rubin diagnostic
should be approximately 1 at convergence.}
}
\value{
a list with the same structure as a single
element of the input. Each element of th return value is a
logical indicating whether the chain has converged
}
\description{
Check convergence of MCMC chains for
all parameters
}
\author{
Sangeeta Bhatia
}
