% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{compute_p_accept_move_E}
\alias{compute_p_accept_move_E}
\title{Computes acceptance probability for a move of the error indicator E}
\usage{
compute_p_accept_move_E(
  direction = c("E0_to_E1", "E1_to_E0"),
  i,
  group_idx,
  date_idx,
  curr_aug_dat,
  proposed_aug_dat,
  theta,
  obs_dat,
  index_dates,
  range_dates = NULL
)
}
\arguments{
\item{direction}{A character string, either "E0_to_E1" or "E1_to_E0",
specifying the direction of the move.}

\item{i}{Index of the individual for whom the move is being made.}

\item{group_idx}{Index of the group for the move.}

\item{date_idx}{Index of the date for the move.}

\item{curr_aug_dat}{The current augmented data before the proposal.}

\item{proposed_aug_dat}{The proposed augmented data after the move.}

\item{theta}{List of model parameters.}

\item{obs_dat}{A list of observed data.}

\item{index_dates}{A list defining the delays for each group.}

\item{range_dates}{A vector containing the range of dates in `obs_dat`.
If NULL, it will be computed automatically.}
}
\value{
A vector of length 2. The first element is the log posterior
  difference (proposed - current), and the second is the log proposal
  correction factor.
}
\description{
Calculates the log acceptance probability for an MCMC
move that flips an error indicator from E=0 to E=1 or vice-versa. It
computes both the log posterior ratio and the proposal correction factor
for the asymmetric move.
}
