% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{compute_p_accept_move_from_E0_to_E1}
\alias{compute_p_accept_move_from_E0_to_E1}
\title{Computes acceptance probability for a move from E = 0 (date observed
 correctly) to E = 1 (date observed incorrectly).}
\usage{
compute_p_accept_move_from_E0_to_E1(
  i,
  group_idx,
  date_idx,
  curr_aug_dat,
  proposed_aug_dat,
  theta,
  obs_dat,
  hyperparameters,
  index_dates,
  range_dates = NULL
)
}
\arguments{
\item{i}{Index of individual(s) for whom augmented data should be moved.}

\item{group_idx}{Index of the group for whom augmented data should be moved.}

\item{date_idx}{Index of the date which should be moved.}

\item{curr_aug_dat}{The current augmented data.}

\item{proposed_aug_dat}{Proposed augmented data.}

\item{theta}{List of parameters; see details.}

\item{obs_dat}{A list of observed data, in the format of the first element
(called \code{obs_dat}) in the list returned by \code{\link{simul_obs_dat}}.}

\item{hyperparameters}{A list of hyperparameters: see details.}

\item{index_dates}{A list containing indications on which delays to consider
in the estimation, see details.}

\item{range_dates}{A vector containing the range of dates in \code{obs_dat}.
If NULL, will be computed automatically.

@return Vector of length 2. The first element is the difference in log
 posterior between the proposed and current augmented data. The second
  is the log proposal correction factor needed to adjust for asymmetry in
   the proposal distribution.}
}
\description{
Computes acceptance probability for a move from E = 0 (date observed
 correctly) to E = 1 (date observed incorrectly).
}
