% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{compute_p_accept_move_from_E1_to_E0}
\alias{compute_p_accept_move_from_E1_to_E0}
\title{Calculates the log acceptance probability for an MCMC move that changes the
 error indicator from E = 1 (observed with error) to E = 0 (observed without
  error) and sets the true date to the observed date.}
\usage{
compute_p_accept_move_from_E1_to_E0(
  i,
  group_idx,
  date_idx,
  curr_aug_dat,
  proposed_aug_dat,
  theta,
  obs_dat,
  hyperparameters,
  index_dates,
  range_dates
)
}
\arguments{
\item{i}{Index of individual(s) for whom augmented data should be moved.}

\item{group_idx}{Index of the group for whom augmented data should be moved.}

\item{date_idx}{Index of the date which should be moved.}

\item{curr_aug_dat}{The current augmented data.}

\item{proposed_aug_dat}{The proposed augmented data.}

\item{theta}{List of parameters; see details.}

\item{obs_dat}{A list of observed data, in the format of the first element
(called \code{obs_dat}) in the list returned by \code{\link{simul_obs_dat}}.}

\item{hyperparameters}{A list of hyperparameters: see details.}

\item{index_dates}{A list containing indications on which delays to consider
in the estimation, see details.}

\item{range_dates}{A vector containing the range of dates in \code{obs_dat}.
If NULL, will be computed automatically.

@return A vector of length 2. The first element is the log posterior
 difference (proposed - current), and the second element is the proposal
  asymmetry correction.}
}
\description{
Calculates the log acceptance probability for an MCMC move that changes the
 error indicator from E = 1 (observed with error) to E = 0 (observed without
  error) and sets the true date to the observed date.
}
