% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{discr_gamma_sample}
\alias{discr_gamma_sample}
\title{Sample from the Discrete Gamma Distribution}
\usage{
discr_gamma_sample(n, mu, cv)
}
\arguments{
\item{n}{Number of samples.}

\item{mu}{Mean of the gamma.}

\item{cv}{Coefficient of variation.}
}
\value{
Integer vector of samples.
}
\description{
Sample from the Discrete Gamma Distribution
}
