% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{DiscrGamma}
\alias{DiscrGamma}
\title{Computes the probability mass of the discrete gamma distribution at integer k.}
\usage{
DiscrGamma(k, mu, cv = NULL, sigma = mu * cv, log = TRUE)
}
\arguments{
\item{k}{Integer (or vector of integers).}

\item{mu}{Mean of the gamma distribution.}

\item{cv}{Coefficient of variation.}

\item{sigma}{Optional standard deviation. Defaults to mu * cv.}

\item{log}{Logical; if TRUE, returns log-probabilities.}
}
\value{
Probability or log-probability mass at k.
}
\description{
Computes the probability mass of the discrete gamma distribution at integer k.
}
