% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{find_params_beta}
\alias{find_params_beta}
\title{Finds the two shape parameters of the beta distribution, given a mean and
 variance}
\usage{
find_params_beta(mean, var)
}
\arguments{
\item{mean}{Mean of the beta distribution}

\item{var}{Variance of the beta distribution}
}
\value{
A vector containing the two shape parameters of the beta
 distribution.
}
\description{
Finds the two shape parameters of the beta distribution, given a mean and
 variance
}
\examples{
param_beta <- find_params_beta(0.1, 0.05)
sample_beta <- rbeta(1000, param_beta[1], param_beta[2])
mean(sample_beta) # compare to 0.1
var(sample_beta) # compare to 0.05
}
