% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{get_param_posterior_estimates}
\alias{get_param_posterior_estimates}
\title{Computes posterior estimates of parameters from the MCMC chain}
\usage{
get_param_posterior_estimates(
  MCMCres,
  central = c("median", "mean"),
  CrI = 0.95,
  theta_true = NULL,
  plot = TRUE,
  group_labels = NULL,
  date_labels = NULL
)
}
\arguments{
\item{MCMCres}{Output from \code{\link{RunMCMC}}.}

\item{central}{A character string specifying what the central estimate
should be (either \code{"median"} or \code{"mean"} posterior)}

\item{CrI}{A scalar in [0;1] used to compute the posterior credible
intervals. For 95\% credible intervals, use CrI=0.95.}

\item{theta_true}{Optional list of true parameter values to compare the
 output to. If not \code{NULL}, this should contain:
\itemize{
 \item{\code{mu}: A list of length
  \code{n_groups = length(MCMCres$aug_dat_chain[[1]]$D)}. Each element of
   \code{mu} should be a scalar or vector giving the mean delay(s) in that
    group.}
 \item{\code{CV}: A list of length \code{n_groups}. Each element of
  \code{CV} should be a scalar or vector giving the coefficient of variation
   of the delay(s) in that group.}
 \item{\code{zeta}: A scalar in [0;1] giving the probability that, if a
  data point is not missing, it is recorded with error.}
}
The posterior distributions of parameters are then plotted together with
 \code{theta_true}.}

\item{plot}{A boolean specifying whether to generate boxplots of the
posterior estimates.}

\item{group_labels}{A character vector of length equal to the number of
groups. Each element gives the label corresponding to a group index e.g.
 \code{c("Community-alive", "Hospitalised-dead")}.}

\item{date_labels}{A list of character vectors giving the names of each date
column per group. This is used to automatically generate delay labels.}
}
\value{
A list with posterior estimates and optional plots:
\itemize{
 \item{\code{logpost}: A vector of three values, containing the central
  log-posterior estimate (mean or median), lower bound of the credible
   interval and upper bound of the credible interval.}
 \item{\code{theta}: A named list giving posterior parameter estimates for:
 \itemize{
 \item{\code{mu}: A list of length equal to the number of groups. Each
  element is a matrix with 3 rows giving the posterior mean/median of the
    delay(s) in that group (1st row = central posterior estimate, 2nd and
     3rd rows = lower and upper credible interval bounds).}
 \item{\code{CV}: Same structure as \code{mu}, giving posterior estimates
  of the coefficient of variation for each delay.}
 \item{\code{zeta}: A vector of three values in [0;1] giving the posterior
  estimate of the probability that, if a data point is not missing, it is
   recorded with error (1st value = central posterior estimate, 2nd and 3rd
    values = lower and upper credible interval bounds).}
 }}
}
If \code{plot = TRUE}, a summary plot of the posterior distributions is also
 displayed. This includes boxplots for the log-posterior, zeta, and the mu
  and cv of delays for each group.
}
\description{
Computes posterior estimates of parameters from the MCMC chain
}
\examples{
# Simulate data to use
n_groups <- 4
index_dates <- list(
  matrix(c(1, 2), nrow = 2),
  cbind(c(1, 2), c(1, 3)),
  cbind(c(1, 2), c(2, 3),c(1, 4)),
  cbind(c(1, 2), c(2, 3), c(1, 4))
)
theta <- list(
  mu = list(5, c(6, 7), c(8, 9, 10), c(11, 12, 13)),
  CV = list(0.5, c(0.5, 0.5), c(0.5, 0.5, 0.5), c(0.5, 0.5, 0.5)),
  prop_missing_data = 0.2,
  zeta = 0.05
 )

n_per_group <- rep(10, n_groups)
range_dates <- c(0, 30)

simul_dat <- simul_true_data(theta, n_per_group, range_dates, index_dates,
                             simul_error = TRUE)
obs_dat <- simul_dat$obs_dat

# Set up hyperparameters
hyperparameters <- list(
    shape1_prob_error = 3,
    shape2_prob_error = 12,
    mean_mean_delay = 10,
    mean_CV_delay = 10)

# Set up MCMC
MCMC_settings <- list(
moves_switch = list(D_on = TRUE, E_on = TRUE, swapE_on = TRUE,
                    mu_on = TRUE, CV_on = TRUE, zeta_on = TRUE),
  moves_options = list(
    fraction_Di_to_update = 1 / 10,
    move_D_by_groups_of_size = 1,
    fraction_Ei_to_update = 1 / 10,
    sdlog_mu = list(
      0.05,
      c(0.15, 0.15),
      c(0.15, 0.15, 0.15),
      c(0.25, 0.25, 0.25)
    ),
    sdlog_CV = list(
      0.25, c(0.25, 0.25), c(0.25, 0.25, 0.25), c(0.25, 0.25, 0.25))
  ),
  init_options = list(
    mindelay = 0,
    maxdelay = 20
  ),
  chain_properties = list(
    n_iter = 500,
    burnin = 50,
    record_every = 10
  )
)

# Run MCMC
MCMC_result <- RunMCMC(obs_dat,
                       MCMC_settings,
                       hyperparameters,
                       index_dates)

# Get summary
get_param_posterior_estimates(MCMCres = MCMC_result,
                              central = "mean",
                              CrI = 0.95,
                              theta_true = NULL,
                              plot = TRUE,
                              group_labels = c("Community-alive",
                                               "Community-dead",
                                               "Hospitalised-alive",
                                               "Hospitalised-dead"),
                              date_labels = list(
                                c("Onset", "Report"),
                                c("Onset", "Death", "Report"),
                                c("Onset", "Hosp", "Disch", "Report"),
                                c("Onset", "Hosp", "Death", "Report")
                              ))


}
\seealso{
\code{\link{RunMCMC}}
}
