% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{identify_affected_delays}
\alias{identify_affected_delays}
\title{Identify delays affected by date changes}
\usage{
identify_affected_delays(index_dates_group, all_changed_dates)
}
\arguments{
\item{index_dates_group}{Index dates defining the delays for a group, where
each column is a delay and the rows are the start/end date indices}

\item{all_changed_dates}{Unique numeric vector containing the indices of all
dates that have been moved or changed e.g. from E0 to E1, from E1 to E0, and
resampled if they were missing}
}
\value{
Unique vector of the indices of all delays affected by the changed
 dates
}
\description{
Identify delays affected by date changes
}
\examples{
# Example using 3 delays (1->2, 2->3, 1->4)
index_dates_group <- matrix(c(1, 2, 1,
                              2, 3, 4), nrow = 2, byrow = TRUE)

# 1. Change two dates in a hypothetical move
changed_dates <- c(2, 4)

# These dates are involved in all 3 delays
identify_affected_delays(index_dates_group, changed_dates)

# 2. Change one date in a hypothetical move
changed_dates <- 1

# Date 1 is only involved in delay indices 1 and 3
identify_affected_delays(index_dates_group, changed_dates)


}
