% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitMCMC.R
\name{initialise_theta_from_aug_dat}
\alias{initialise_theta_from_aug_dat}
\title{Initialises parameters data on augmented data (for the MCMC)}
\usage{
initialise_theta_from_aug_dat(aug_dat, index_dates, zeta_init = 0.1)
}
\arguments{
\item{aug_dat}{A list of data, in the format returned by
\code{\link{simul_true_data}}.}

\item{index_dates}{A list containing indications on which delays to consider
in the simulation, same as in \code{\link{simul_true_data}}.}

\item{zeta_init}{A scalar giving the value zeta should be initialised to.}
}
\value{
A list containing:
\itemize{
 \item{\code{mu}}{: A list of length \code{n_groups} (the number of groups
  to be simulated data). Each element of \code{mu} should be a scalar of
   vector giving the mean delay(s) to use for simulation of dates in that
   group.}
 \item{\code{CV}}{: A list of length \code{n_groups}. Each element of
  \code{CV} should be a scalar of vector giving the coefficient of variation
   of the delay(s) to use for simulation of dates in that group.}
 \item{\code{zeta}}{: A scalar in [0;1] giving the probability that, if a
  data point is not missing, it is recorded with error.}
}
}
\description{
Initialises parameters data on augmented data (for the MCMC)
}
\examples{
### Number of groups of individuals to simulate ###
n_groups <- 2
### Number of dates to simulate for each group ###
n_dates <- c(2, 3)
### Setting up the parameters for the simulation ###
theta <- list()
theta$mu <- list(5, c(10, 15)) # mean delays, for each group
theta$CV <- list(0.5, c(0.5, 0.5)) # coefficient of variation of these delays
theta$prop_missing_data <- 0.25 # probability of data missing in observations
theta$zeta <- 0.05 # probability that, when not missing, the date is recorded
 with error
### Number of individuals to simulate in each group ###
n_per_group <- rep(10, n_groups)
### Range of dates in which to draw the first set of dates for each group ###
range_dates <- date_to_int(c(as.Date("01/01/2014", "\%d/\%m/\%Y"),
 as.Date("01/01/2015", "\%d/\%m/\%Y")))
### Which delays to use to simulate subsequent dates from the first, in each
 group? ###
index_dates <- list(matrix(c(1, 2), nrow=2), cbind(c(1, 2), c(1, 3)))
### Perform the simulation ###
D <- simul_true_data(theta, n_per_group, range_dates, index_dates)
observed_D <- simul_obs_dat(D$true_dat, theta, range_dates,
 remove_allNA_indiv=TRUE)
### Initialise augmented data first ###
MCMC_settings <- list(init_options=list(mindelay=0, maxdelay=100))
aug_dat <- initialise_aug_data(observed_D$obs_dat, index_dates,
 MCMC_settings)
### Now initialise parameters based on the augmented data above ###
theta <- initialise_theta_from_aug_dat(aug_dat, index_dates)
}
