% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodPrior.R
\name{LL_delays_term}
\alias{LL_delays_term}
\title{Compute total delay likelihood across all individuals and groups}
\usage{
LL_delays_term(aug_dat, theta, obs_dat, index_dates, Delta = NULL)
}
\description{
Compute total delay likelihood across all individuals and groups
}
