% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodPrior.R
\name{LL_delays_term_by_group_delay_and_indiv}
\alias{LL_delays_term_by_group_delay_and_indiv}
\title{Compute delay likelihood for individual delay observations}
\usage{
LL_delays_term_by_group_delay_and_indiv(
  aug_dat,
  theta,
  obs_dat,
  group_idx,
  delay_idx,
  indiv_idx,
  index_dates,
  Delta = NULL
)
}
\description{
Compute delay likelihood for individual delay observations
}
