% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodPrior.R
\name{LL_observation_term}
\alias{LL_observation_term}
\title{Compute total observation likelihood over all individuals, dates, and groups}
\usage{
LL_observation_term(aug_dat, obs_dat, range_dates = NULL)
}
\description{
Compute total observation likelihood over all individuals, dates, and groups
}
