% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodPrior.R
\name{LL_total}
\alias{LL_total}
\title{Compute full likelihood (observation + delay + error)}
\usage{
LL_total(aug_dat, theta, obs_dat, index_dates, range_dates = NULL)
}
\description{
Compute full likelihood (observation + delay + error)
}
