% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodPrior.R
\name{lprior_params_delay}
\alias{lprior_params_delay}
\title{Log prior for mu or CV (Exponential prior)}
\usage{
lprior_params_delay(what = c("mu", "CV"), theta, hyperparameters)
}
\description{
Log prior for mu or CV (Exponential prior)
}
