% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodPrior.R
\name{lprior_prob_error}
\alias{lprior_prob_error}
\title{Log prior for error probability parameter zeta (Beta prior)}
\usage{
lprior_prob_error(theta, hyperparameters)
}
\description{
Log prior for error probability parameter zeta (Beta prior)
}
