% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{move_zeta_gibbs}
\alias{move_zeta_gibbs}
\title{Performs one iteration of an MCMC move for the parameter zeta (probability
 of a data being recorded erroneously, given it is recorded)}
\usage{
move_zeta_gibbs(aug_dat, curr_theta, obs_dat, hyperparameters)
}
\arguments{
\item{aug_dat}{The augmented data; a list of observed data, in the format
returned by \code{\link{simul_true_data}}.}

\item{curr_theta}{The current list of parameters; see details.}

\item{obs_dat}{A list of observed data, in the format of the first element
(called \code{obs_dat}) in the list returned by \code{\link{simul_obs_dat}}.}

\item{hyperparameters}{A list of hyperparameters: see details.}
}
\value{
A list of two elements:
 \itemize{
 \item{\code{new_theta}: Same as \code{curr_theta} but where
  \code{curr_theta$zeta} has been updated}
 \item{\code{accept}: A scalar with value 1 (as we are using a Gibbs
  sampler the move is always accepted)}
}
}
\description{
Performs one iteration of an MCMC move for the parameter zeta (probability
 of a data being recorded erroneously, given it is recorded)
}
\details{
\code{curr_theta} should be a list containing:
\itemize{
 \item{\code{mu}}{: A list of length \code{n_groups} (the number of groups
  to be simulated data). Each element of \code{mu} should be a scalar of
   vector giving the mean delay(s) to use for simulation of dates in that
    group.}
 \item{\code{CV}}{: A list of length \code{n_groups}. Each element of
  \code{CV} should be a scalar of vector giving the coefficient o variation
   of the delay(s) to use for simulation of dates in that group.}
 \item{\code{zeta}}{: A scalar in [0;1] giving the probability that, if a
  data point is not missing, it is recorded with error.}
}
\code{hyperparameters} should be a list containing:
\itemize{
 \item{\code{shape1_prob_error}}{: A scalar giving the first shape parameter
  for the beta prior used for parameter \code{theta$zeta}}
 \item{\code{shape2_prob_error}}{: A scalar giving the second shape
  parameter for the beta prior used for parameter \code{theta$zeta}}
 \item{\code{mean_mean_delay}}{: A scalar giving the mean of the exponential
  prior used for parameter \code{theta$mu}}
 \item{\code{mean_CV_delay}}{: A scalar giving the mean of the exponential
  prior used for parameter \code{theta$CV}}
}

The function performs the move, using a Gibbs sampler.
A new value of parameter zeta is drawn from its marginal posterior
 distribution, that is a beta distribution with parameters:
\itemize{
 \item{\code{first shape parameter}}{: Equal to
  \code{hyperparameters$shape1_prob_error} + number_of_errors, where
   number_of_errors is the number of data points recorded with errors}
 \item{\code{second shape parameter}}{: Equal to
  \code{hyperparameters$shape2_prob_error} + number_of_recorded_dates -
   number_of_errors, where number_of_recorded_dates is the number of data
    points which are not missing, and number_of_errors is the number of data
     points recorded with errors}
}
}
\examples{
### TO WRITE OR ALTERNATIVELY REFER TO VIGNETTE TO BE WRITTEN ###
}
