% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{perform_E0_to_E1_swap}
\alias{perform_E0_to_E1_swap}
\title{Perform the E=0 to E=1 swap for a set of dates}
\usage{
perform_E0_to_E1_swap(
  i,
  group_idx,
  date_idx_E0_to_E1,
  current_aug_dat,
  theta,
  obs_dat,
  hyperparameters,
  index_dates,
  range_dates
)
}
\arguments{
\item{i}{Index of the individual.}

\item{group_idx}{Index of the group.}

\item{date_idx_E0_to_E1}{Numeric vector of column indices for the dates
that are `E=0` and will be moved.}

\item{current_aug_dat}{Augmented data list after the E=1 to E=0 swap
has already been performed in the preceding step.}

\item{theta}{List of model parameters (`mu`, `CV`, `zeta`).}

\item{obs_dat}{List of the observed data.}

\item{hyperparameters}{List of the model's hyperparameters.}

\item{index_dates}{List defining the delays for each group.}

\item{range_dates}{Vector containing the overall date range.}
}
\value{
A list containing two elements:
\itemise{
  \item{`proposed_aug_dat`}: Augmented data with the relevant dates
    updated to the new `E=1` state.
  \item{`correction_factor`}: Sum of the log-probabilities for the
    proposal of these dates.
}
}
\description{
Second sequential step within `swap_Ei`. It takes all dates for a given
individual that are currently marked as correct (`E=0`) and proposes a new
state where they are marked as an error (`E=1`).
}
\details{
This is a non-deterministic proposal. For each date being moved, the error
indicator `E` is set to `1`, and a new plausible true date `D` is sampled
via `propose_move_from_E0_to_E1`. This sampling is conditional on the other
dates, including those updated in the previous step of the swap.

This function also calculates and accumulates the log-probability
of making these sequential proposals. This is done by calling
`compute_p_accept_move_E` for each move and summing the resulting
correction factors.
}
\seealso{
`swap_Ei`, `propose_move_from_E0_to_E1`, `compute_p_accept_move_E`
}
