% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateHandWrittenDates.R
\name{plot_handwritten_date}
\alias{plot_handwritten_date}
\title{Plots a handwritten date}
\usage{
plot_handwritten_date(
  date = as.Date("01/01/2017", format = "\%d/\%m/\%Y"),
  d = load_mnist(download_if_missing = TRUE)
)
}
\arguments{
\item{date}{A date}

\item{d}{an object of class mnist containing a database of handwritten digits (see \code{rmnist::load_mnist}).}
}
\description{
Plots a handwritten date
}
\examples{
plot_handwritten_date(Sys.Date()) # print today's date
}
