% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{propose_move_from_E0_to_E1}
\alias{propose_move_from_E0_to_E1}
\title{Propose a new value for the true date when transitioning the error indicator
 from 0 (no error) to 1 (observed with error), ensuring the new date is
  different from the observed date.}
\usage{
propose_move_from_E0_to_E1(
  i,
  group_idx,
  date_idx,
  curr_aug_dat,
  theta,
  obs_dat,
  hyperparameters,
  index_dates,
  range_dates = NULL
)
}
\arguments{
\item{i}{Index of individual(s) for whom augmented data should be moved.}

\item{group_idx}{Index of the group for whom augmented data should be moved.}

\item{date_idx}{Index of the date which should be moved.}

\item{curr_aug_dat}{The current augmented data.}

\item{theta}{List of parameters, including mu, CV and zeta.}

\item{obs_dat}{A list of observed data, in the format of the first element
(called \code{obs_dat}) in the list returned by \code{\link{simul_obs_dat}}.}

\item{hyperparameters}{A list of hyperparameters.}

\item{index_dates}{A list containing the delays defined for each group.}

\item{range_dates}{A vector containing the range of dates in \code{obs_dat}.
If NULL, will be computed automatically.}
}
\value{
A vector of proposed true dates that differ from the observed dates.
}
\description{
Propose a new value for the true date when transitioning the error indicator
 from 0 (no error) to 1 (observed with error), ensuring the new date is
  different from the observed date.
}
