% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{propose_move_from_E1_to_E0}
\alias{propose_move_from_E1_to_E0}
\title{Propose the observed date when transitioning the error indicator from 1
 (observed with error) to 0 (observed with no error). This move assumes the
 true date was correctly recorded and simply sets D to the observed date.}
\usage{
propose_move_from_E1_to_E0(i, group_idx, date_idx, obs_dat)
}
\arguments{
\item{i}{Index of individual(s) for whom augmented data should be moved.}

\item{group_idx}{Index of the group for whom augmented data should be moved.}

\item{date_idx}{Index of the date which should be moved.}

\item{obs_dat}{A list of observed data, in the format of the first element
(called \code{obs_dat}) in the list returned by \code{\link{simul_obs_dat}}.}
}
\value{
Proposed true date, which is the same as the observed date.
}
\description{
Propose the observed date when transitioning the error indicator from 1
 (observed with error) to 0 (observed with no error). This move assumes the
 true date was correctly recorded and simply sets D to the observed date.
}
