% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{resample_missing_dates}
\alias{resample_missing_dates}
\title{Resample missing dates}
\usage{
resample_missing_dates(
  i,
  group_idx,
  date_idx_resample,
  current_aug_dat,
  theta,
  obs_dat,
  hyperparameters,
  index_dates,
  range_dates
)
}
\arguments{
\item{i}{Index of the individual for whom the dates are being resampled}

\item{group_idx}{Index of the group for the individual}

\item{date_idx_resample}{Numeric vector of column indixes for the missing
dates (`E=-1`) that need to be resampled.}

\item{current_aug_dat}{Augmented data list after the E=0/E=1 swaps that have
already been performed in the preceding steps.}

\item{theta}{List of model parameters (`mu`, `CV`, `zeta`)}

\item{obs_dat}{List of observed data}

\item{hyperparameters}{List of model hyperparameters}

\item{index_dates}{List defining the delays for each group}

\item{range_dates}{Vector containing the overall range of dates to consider}
}
\value{
List containing two elements:
\itemise{
 \item{`proposed_aug_dat`}: Augmented data with missing dates resampled to
   be consistent with the other proposed changes.
 \item{`correction_factor`}: Sum of the log-probabilities for the proposal
   of the missing dates.
}
}
\description{
Resamples any missing dates (`E=-1`) within the `swap_Ei` function to ensure
they remain consistent with the newly proposed dates from the preceding
steps of the swap.
}
\details{
After the error indicators (`E=0` and `E=1`) have been swapped, this function
iterates through dates marked as missing (`E=-1`) for a given individual. For
each missing date it calls `propose_new_delay()` to generate a new date that
is conditional on the already updated dates. It also calculates the log-
probability of making these proposals.
}
