% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{reverse_E0_to_E1_swap}
\alias{reverse_E0_to_E1_swap}
\title{Calculate reverse proposal probability for E=0 to E=1 swaps}
\usage{
reverse_E0_to_E1_swap(
  i,
  group_idx,
  date_idx_E0_to_E1,
  current_reverted_dat,
  original_dat,
  theta,
  obs_dat,
  index_dates,
  range_dates
)
}
\arguments{
\item{i}{Index of the individual.}

\item{group_idx}{Index of the group.}

\item{date_idx_E0_to_E1}{Numeric vector of column indices for the dates
that were `E=0` in the original data.}

\item{current_reverted_dat}{Augmented data list after the missing dates
have already been reverted in the preceding step.}

\item{original_dat}{Augmented data list before any proposals were made.}

\item{theta}{List of model parameters (`mu`, `CV`, `zeta`).}

\item{obs_dat}{List of the observed data.}

\item{index_dates}{List defining the delays for each group.}

\item{range_dates}{Vector containing the overall range of dates.}
}
\value{
A list containing two elements:
\itemize{
  \item{`reverted_aug_dat`}: Augmented data with the relevant dates
    reverted back to their original `E=0` state.
  \item{`correction_factor`}: Sum of the log-probabilities for this
    part of the reverse proposal.
}
}
\description{
Sequential step within `swap_Ei`. It performs the reverse of the
`perform_E0_to_E1_swap` step. It reverts the error indicators (`E`) and
dates (`D`) for the relevant dates back to their original state and
calculates the correction factor for this reverse move.
}
\details{
Takes the partially reverted data from the previous step
(`reverse_resample_missing_dates`) and continues the reversal process. It
iterates backwards (`rev()`) through the dates that were originally `E=0`
and were swapped to `E=1` in the forward move.

For each date, it reverts both the `E` and `D` values to their state in the
original data. It then calculates the log-probability of the forward move
from this newly reverted state back to the state before the revert. This
value is accumulated to create the total correction factor for this part of
the reverse path.
}
\seealso{
`perform_E0_to_E1_swap`, `compute_p_accept_move_E`
}
