% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{reverse_resample_missing_dates}
\alias{reverse_resample_missing_dates}
\title{Calculate reverse proposal probability for resampled missing dates}
\usage{
reverse_resample_missing_dates(
  i,
  group_idx,
  date_idx_resample,
  final_proposed_dat,
  original_dat,
  theta,
  index_dates
)
}
\arguments{
\item{i}{Index of the individual.}

\item{group_idx}{Index of the group.}

\item{date_idx_resample}{Numeric vector of column indices for the dates
that were resampled in the forward move.}

\item{final_proposed_dat}{Augmented data list after the full forward
proposal has been completed.}

\item{original_dat}{Augmented data list before any proposals were made.}

\item{theta}{List of model parameters (`mu`, `CV`, `zeta`).}

\item{index_dates}{List defining the delays for each group.}
}
\value{
A list containing two elements:
\itemise{
  \item{`reverted_aug_dat`}: Augmented data with the resampled dates
    reverted back to their original state.
  \item{`correction_factor`}: Sum of the log-probabilities for the
    reverse proposal of the missing dates.
}
}
\description{
Sequential step within `swap_Ei`. It performs the reverse of the
`resample_missing_dates` step. It reverts the dates for missing indices
back to their original values and calculates the correction factor for this
reverse move.
}
\seealso{
`resample_missing_dates`
}
