% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulFunctions.R
\name{simul_true_data}
\alias{simul_true_data}
\title{Simulates a dataset with true event dates, and optionally adds missingness
 and errors.}
\usage{
simul_true_data(
  theta,
  n_per_group,
  range_dates,
  index_dates,
  simul_error = FALSE,
  remove_allNA_indiv = TRUE
)
}
\arguments{
\item{theta}{List of parameters for delay distributions and data errors (see
details).}

\item{n_per_group}{Vector containing the number of individuals to simulate
per group (i.e., community-alive, hospitalised-dead, etc.).}

\item{range_dates}{Vector of two integers: the range to draw the initial date
from (uniformly).}

\item{index_dates}{A list defining how each group's date delays are simulated
(see details).}

\item{simul_error}{Boolean. If TRUE, simulate missing and erroneous data.}

\item{remove_allNA_indiv}{Boolean. If TRUE, remove rows for individuals with
all dates missing (only used if \code{simul_error = TRUE}).}
}
\value{
A list with three elements: 
 \itemize{
   \item{\code{true_dat}: List of matrices containing the simulated true
    dates as integers for each group. Each row corresponds to an individual;
     columns correspond to event dates (e.g., onset, report) as defined in
      \code{index_dates}.}
   \item{\code{obs_dat}: Same structure as \code{true_dat} but with simulated
    errors and missing dates (NULL if \code{simul_error = FALSE}).}
    \item{\code{E}: Error indicator matrices with the same structure as
     \code{true_dat} and \code{obs_dat}, where each element indicates: -1 for
      missing date, 1 for erroneous data, or 0 for correctly recorded data. E
       will be NULL if \code{simul_error = FALSE}.}
 }
}
\description{
Simulates a dataset with true event dates, and optionally adds missingness
 and errors.
}
\details{
\code{theta} should be a list containing:
\itemize{
 \item{\code{mu}: A list of length \code{n_groups} (the number of groups
  to be simulated). Each element should be a scalar or vector giving
   the mean delay(s) used to simulate dates in that group.}
 \item{\code{CV}: A list of length \code{n_groups}. Each element of \code{CV}
  should be a scalar or vector giving the coefficient of variation for the
   delay(s) used to simulate dates in that group.}
 \item{\code{prop_missing_data} (only required if \code{simul_error = TRUE}):
  A scalar in [0,1] giving the probability that a date is missing.}
 \item{\code{zeta} (only required if \code{simul_error = TRUE}): A scalar
  in [0,1] giving the probability that, if a date is not missing, it is
   recorded with error.}
}

\code{index_dates} should be a list of length \code{n_groups}. Each element
 should be a matrix with 2 rows and columns corresponding to the delays of
  interest for that group. For each column:
  \itemize{
   \item Row 1 gives the index of the origin date.
   \item Row 2 gives the index of the destination date.
   }
The number of columns in \code{index_dates[[k]]} should match the length of
 \code{theta$mu[[k]]} and \code{theta$CV[[k]]}.

If index_dates[[k]] has two columns containing respectively c(1, 2) and
 c(1, 3), this indicates that theta$mu[[k]] and theta$CV[[k]] are
  respectively the mean and coefficient of variation of two delays: the
   first delay being between date 1 and date 2, and the second being between
    date 1 and date 3.

In the simulation, date 1 will be drawn uniformly within \code{range_dates}. 
Then date 2 will be drawn as date 1 + a delay drawn from a discretised gamma
 distribution with mean theta$mu[[k]][1] and CV theta$CV[[k]][1]. 
Finally, date 3 will be drawn as date 1 + a delay drawn from a discretised
 gamma distribution with mean theta$mu[[k]][1] and CV theta$CV[[k]][1].
 
 @seealso
   [simul_obs_dat()] for simulating missing and erroneous data,
   [date_to_int()] for converting dates to integers corresponding to the
    number of days from a given origin,
   [int_to_date()] for converting integers to dates based on a given origin
    from which the integer counts the number of days.
}
\examples{
# Number of groups of individuals to simulate
n_groups <- 2

# Number of dates to simulate for each group
n_dates <- c(2, 3)

# Set up the parameters for the simulation
theta <- list()
theta$mu <- list(5, c(10, 15)) # mean delays, for each group
theta$CV <- list(0.5, c(0.5, 0.5)) # coefficient of variation of delays

# Number of individuals to simulate in each group
n_per_group <- rep(10, n_groups)

# Range of dates in which to draw first set of dates for each group
range_dates <- date_to_int(c(as.Date("01/01/2014", "\%d/\%m/\%Y"),
                             as.Date("01/01/2015", "\%d/\%m/\%Y")))

# Delays to use to simulate subsequent dates from the first, in each group
index_dates <- list(matrix(c(1, 2), nrow = 2), cbind(c(1, 2), c(1, 3)))

# Perform the simulation
D <- simul_true_data(theta, n_per_group, range_dates, index_dates)
}
