% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{split_chain_in_two}
\alias{split_chain_in_two}
\title{Splits a single MCMC chain into two}
\usage{
split_chain_in_two(chain)
}
\arguments{
\item{chain}{numeric vector, samples from MCMC chain.}
}
\value{
\code{mcmc.list} from {coda}
}
\description{
Splits a single MCMC chain into two
}
\details{
This is a utility function
and should not have to be called directly
by the user.
}
