% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{compute_autocorr}
\alias{compute_autocorr}
\title{Compute autocorrelation for each parameter of the MCMC chains}
\usage{
compute_autocorr(MCMCres)
}
\arguments{
\item{MCMCres}{The output of function \code{\link{RunMCMC}}.}
}
\value{
A list of results of autocorrelation results (obtained from the function \code{\link{acf}}).
}
\description{
Compute autocorrelation for each parameter of the MCMC chains
}
\examples{
### TO WRITE OR ALTERNATIVELY REFER TO VIGNETTE TO BE WRITTEN ###
}
