% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{compute_correlations_mu_CV}
\alias{compute_correlations_mu_CV}
\title{Compute correlation between the MCMC chains of mean and CV of each delay}
\usage{
compute_correlations_mu_CV(MCMCres, index_dates, plot = TRUE)
}
\arguments{
\item{MCMCres}{The output of function \code{\link{RunMCMC}}.}

\item{index_dates}{A list containing indications on which delays to consider in the estimation, see details.}

\item{plot}{A boolean indicating whether to plot the correlations or not}
}
\value{
A list of results of correlation test (obtained from the function \code{\link{cor.test}}) between the posterior mean and the posterior CV for each delay.
}
\description{
Compute correlation between the MCMC chains of mean and CV of each delay
}
\details{
\code{index_dates} should be a list of length \code{n_groups=length(obs_dat)}. Each element of \code{index_dates} should be a matrix with 2 rows and a number of columns corresponding to the delays of interest for that group. For each column (i.e. each delay), the first row gives the index of the origin date, and the second row gives the index of the destination date. 
The number of columns of index_dates[[k]] should match the length of theta$mu[[k]] and theta$CV[[k]] 

If index_dates[[k]] has two columns containing respectively c(1, 2) and c(1, 3), this indicates that theta$mu[[k]] and theta$CV[[k]] are respectively the mean and coefficient of variation of two delays: the first delay being between date 1 and date 2, and the second being between date 1 and date 3.
}
\examples{
### TO WRITE OR ALTERNATIVELY REFER TO VIGNETTE TO BE WRITTEN ###
}
