% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{compute_correlations_mu_CV}
\alias{compute_correlations_mu_CV}
\title{Compute correlation between the MCMC chains of mean and CV of each delay}
\usage{
compute_correlations_mu_CV(MCMCres, plot = TRUE)
}
\arguments{
\item{MCMCres}{The output of function \code{\link{RunMCMC}}.}

\item{plot}{A boolean indicating whether to plot the correlations or not}
}
\value{
A list of results of correlation test (obtained from the function \code{\link{cor.test}}) between the posterior mean and the posterior CV for each delay.
}
\description{
Compute correlation between the MCMC chains of mean and CV of each delay
}
\examples{
### TO WRITE OR ALTERNATIVELY REFER TO VIGNETTE TO BE WRITTEN ###
}
