% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{compute_delta}
\alias{compute_delta}
\title{Compute relevant delays based on data}
\usage{
compute_delta(D, index_dates)
}
\arguments{
\item{D}{A list of data, in the format of the first element (called \code{true_dat}) in the list returned by \code{\link{simul_true_data}}.}

\item{index_dates}{A list containing indications on which delays to consider in the simulation.}
}
\value{
A list of same length as \code{D}. Each element in the list is a matrix with same number of rows as in \code{D}, but potentially a different number of columns, corresponding to the relevant delays calculated for that group.
}
\description{
Compute relevant delays based on data
}
\details{
\code{index_dates} should be a list of length \code{n_groups=length(D)}. Each element of \code{index_dates} should be a matrix with 2 rows and a number of columns corresponding to the delays of interest for that group.
For each column (i.e. each delay), the first row gives the index of the origin date, and the second row gives the index of the destination date.
}
\examples{
### Number of groups of individuals to simulate ###
n_groups <- 2
### Number of dates to simulate for each group ###
n_dates <- c(2, 3)
### Setting up the parameters for the simulation ###
theta <- list()
theta$mu <- list(5, c(10, 15)) # mean delays, for each group
theta$CV <- list(0.5, c(0.5, 0.5)) # coefficient of variation of these delays
### Number of individuals to simulate in each group ###
n_per_group <- rep(10, n_groups)
### Range of dates in which to draw the first set of dates for each group ###
range_dates <- date_to_int(c(as.Date("01/01/2014", "\%d/\%m/\%Y"), as.Date("01/01/2015", "\%d/\%m/\%Y")))
### Which delays to use to simulate subsequent dates from the first, in each group? ###
index_dates <- list(matrix(c(1, 2), nrow = 2), cbind(c(1, 2), c(1, 3)))
### Perform the simulation ###
D <- simul_true_data(theta, n_per_group, range_dates, index_dates)
### Compute the corresponding delays ###
Delays <- compute_delta(D$true_dat, index_dates)
}
