% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{date_to_int}
\alias{date_to_int}
\title{Convert date to integer corresponding to number of days from a given origin}
\usage{
date_to_int(date, origin = "1970-01-01")
}
\arguments{
\item{date}{\code{Date} object to be converted to integer.}

\item{origin}{The date used as origin for counting days.}
}
\value{
An integer corresponding to the number of days since \code{origin}.
}
\description{
Convert date to integer corresponding to number of days from a given origin
}
\examples{
date_to_int(as.Date("1970-01-01"))
}

