% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{decide_acceptance}
\alias{decide_acceptance}
\title{Metropolis-Hastings accept/reject step}
\usage{
decide_acceptance(log_p_accept, proposed_dat, current_dat)
}
\arguments{
\item{log_p_accept}{Log acceptance probability. Sum of the log posterior
ratio and the log proposal correction factor. The function handles `NA`
values by treating them as `-Inf` and caps positive values at `0`.}

\item{proposed_dat}{Augmented data representing the new, proposed state.
This will be returned if the move is accepted.}

\item{current_dat}{Original state before the proposal. This will be returned
if the move is rejected.}
}
\value{

}
\description{
Helper function to perform accept/reject step of a Metropolis-Hastings move.
It compares the final log posterior acceptance probability to a random draw
from a uniform distribution to decide whether to accept the proposed state
or keep the current state.
}
\examples{
current_state <- list(D = 10)
proposed_state <- list(D = 15)

set.seed(1)
# Expect to accept
decide_acceptance(log_p_accept = 0, proposed_state, current_state)

# Expect to reject
decide_acceptance(log_p_accept = -10, proposed_state, current_state)

}
