% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{find_range}
\alias{find_range}
\title{Find range of dates from a dataset}
\usage{
find_range(obs_dat)
}
\arguments{
\item{obs_dat}{A list of data, in the format of the first element (called \code{obs_dat}) in the list returned by \code{\link{simul_obs_dat}}.}
}
\value{
A vector of two integers coresponding to the range in \code{obs_dat}.
}
\description{
Find range of dates from a dataset
}
\examples{
### Number of groups of individuals to simulate ###
n_groups <- 2
### Number of dates to simulate for each group ###
n_dates <- c(2, 3)
### Setting up the parameters for the simulation ###
theta <- list()
theta$mu <- list(5, c(10, 15)) # mean delays, for each group
theta$CV <- list(0.5, c(0.5, 0.5)) # coefficient of variation of these delays
### Number of individuals to simulate in each group ###
n_per_group <- rep(10, n_groups)
### Range of dates in which to draw the first set of dates for each group ###
range_dates <- date_to_int(c(as.Date("01/01/2014", "\%d/\%m/\%Y"), as.Date("01/01/2015", "\%d/\%m/\%Y")))
### Which delays to use to simulate subsequent dates from the first, in each group? ###
index_dates <- list(matrix(c(1, 2), nrow=2), cbind(c(1, 2), c(1, 3)))
### Perform the simulation ###
D <- simul_true_data(theta, n_per_group, range_dates, index_dates)
### Find the range ###
find_range(D$true_dat)
### Compare with range specified in the first place ### 
range_dates
}

