% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{get_param_posterior_estimates}
\alias{get_param_posterior_estimates}
\title{Computes posterior estimates of parameters from the MCMC chain}
\usage{
get_param_posterior_estimates(MCMCres, index_dates, central = c("median",
  "mean"), CrI = 0.95, theta_true = NULL, plot = TRUE,
  cex.axis = 1)
}
\arguments{
\item{MCMCres}{The output of function \code{\link{RunMCMC}}.}

\item{index_dates}{A list containing indications on which delays to consider in the estimation, see details.}

\item{central}{A character specifying what the central estimate should be (median or mean posterior)}

\item{CrI}{A scalar in [0;1] used to compute the posterior credible intervals. For 95\% credible intervals, use CrI=0.95.}

\item{theta_true}{A list of parameters to which the output chains should be compared. If not \code{NULL}, this should contain:
\itemize{
 \item{\code{mu}}{: A list of length \code{n_groups=length(MCMCres$aug_dat_chain[[1]]$D)}. Each element of \code{mu} should be a scalar or vector giving the mean delay(s) in that group.}
 \item{\code{CV}}{: A list of length \code{n_groups}. Each element of \code{CV} should be a scalar or vector giving the coefficient o variation of the delay(s) in that group.}
 \item{\code{zeta}}{: A scalar in [0;1] giving the probability that, if a data point is not missing, it is recorded with error.}
}
The posterior distributions of parameters are then plotted together with \code{theta_true}.}

\item{plot}{A boolean specifying whether to plot boxplots of the posterior estimates or not}

\item{cex.axis}{A numerical value giving the amount by which x axis labels should be magnified relative to the default.}
}
\value{
A list containing two elements: the posterior estimates of parameters:
\itemize{
 \item{\code{logpost}}{: A vector of three values giving the central log-posterior estimate (first value) and quantiles corresponding to CrI (second and third values). }
 \item{\code{theta}}{: A list giving posterior parameter estimates
 \itemize{
 \item{\code{mu}}{: A list of length \code{n_groups=length(MCMCres$aug_dat_chain[[1]]$D)}. Each element of \code{mu} should be a matrix with 3 rows giving the posterior mean delay(s) in that group (1st row = central posterior estimate, 2nd and 3rd rows = credible interval) .}
 \item{\code{CV}}{: A list of length \code{n_groups}. Each element of \code{CV} should be a matrix with 3 rows giving the posterior CV of the delay(s) in that group (1st row = central posterior estimate, 2nd and 3rd rows = credible interval) .}
 \item{\code{zeta}}{: A vector of three values in [0;1] giving the posterior estimate of the probability that, if a data point is not missing, it is recorded with error (1st value = central posterior estimate, 2nd and 3rd values = credible interval).}
 }
 }
}
}
\description{
Computes posterior estimates of parameters from the MCMC chain
}
\details{
\code{index_dates} should be a list of length \code{n_groups=length(obs_dat)}. Each element of \code{index_dates} should be a matrix with 2 rows and a number of columns corresponding to the delays of interest for that group. For each column (i.e. each delay), the first row gives the index of the origin date, and the second row gives the index of the destination date. 
The number of columns of index_dates[[k]] should match the length of theta$mu[[k]] and theta$CV[[k]] 

If index_dates[[k]] has two columns containing respectively c(1, 2) and c(1, 3), this indicates that theta$mu[[k]] and theta$CV[[k]] are respectively the mean and coefficient of variation of two delays: the first delay being between date 1 and date 2, and the second being between date 1 and date 3.
}
\examples{
### TO WRITE OR ALTERNATIVELY REFER TO VIGNETTE TO BE WRITTEN ###
}
