% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{int_to_date}
\alias{int_to_date}
\title{Convert integer to date, based on a given origin from which the integer counts the number of days}
\usage{
int_to_date(int, origin = "1970-01-01")
}
\arguments{
\item{int}{integer to be converted to \code{Date}.}

\item{origin}{The date used as origin for counting days.}
}
\value{
A \code{Date} object corresponding to \code{int} days after \code{origin}.
}
\description{
Convert integer to date, based on a given origin from which the integer counts the number of days
}
\examples{
int_to_date(365)
}
