% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodPrior.R
\name{LL_error_term}
\alias{LL_error_term}
\title{Compute total log-likelihood for observed errors}
\usage{
LL_error_term(aug_dat, theta, obs_dat)
}
\description{
Compute total log-likelihood for observed errors
}
