% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodPrior.R
\name{LL_observation_term_by_group_delay_and_indiv}
\alias{LL_observation_term_by_group_delay_and_indiv}
\title{Compute log-likelihood of observed dates for a subset of individuals and dates}
\usage{
LL_observation_term_by_group_delay_and_indiv(
  aug_dat,
  theta,
  obs_dat,
  group_idx,
  date_idx,
  indiv_idx,
  range_dates = NULL
)
}
\description{
Calculates the likelihood of the observed dates (\code{obs_dat})
 given the augmented dates (\code{aug_dat}) for a specific group, set of
  individuals, and dates.
}
\details{
Assumes that:
 - if no error (E = 0), true date must equal observed date.
 - If error or missing (E = 1 or -1), observed date is treated as uniformly
  likely over \code{range_dates}.
}
