% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{move_Ei}
\alias{move_Ei}
\title{Performs one iteration of an MCMC move for the augmented data representing
 the indicator of error in observations}
\usage{
move_Ei(
  i,
  group_idx,
  date_idx,
  curr_aug_dat,
  theta,
  obs_dat,
  hyperparameters,
  index_dates,
  range_dates = NULL
)
}
\arguments{
\item{i}{The index of the individual for whom augmented data should be moved}

\item{group_idx}{The index of the group for whom augmented data should be
moved}

\item{date_idx}{The index of the date for which the error indicator which
should be moved}

\item{curr_aug_dat}{The current augmented data; a list of observed data, in
the format returned by \code{\link{simul_true_data}}.}

\item{theta}{List of parameters; see details.}

\item{obs_dat}{A list of observed data, in the format of the first element
(called \code{obs_dat}) in the list returned by \code{\link{simul_obs_dat}}.}

\item{hyperparameters}{A list of hyperparameters: see details.}

\item{index_dates}{A list containing indications on which delays to consider
in the estimation, see details.}

\item{range_dates}{A vector containing the range of dates in \code{obs_dat}.
If NULL, will be computed automatically.}
}
\value{
A list of two elements:
 \itemize{
 \item{\code{new_aug_dat}}{: Same as \code{curr_aug_dat} but where the
  relevant indicators of errors in dates have been updated}
 \item{\code{accept}}{: A scalar with value 1 if the move was accepted and
  0 otherwise}
}
}
\description{
Performs one iteration of an MCMC move for the augmented data representing
 the indicator of error in observations
}
\details{
\code{theta} should be a list containing:
\itemize{
 \item{\code{mu}}{: A list of length \code{n_groups} (the number of groups
  to be simulated data). Each element of \code{mu} should be a scalar of
   vector giving the mean delay(s) to use for simulation of dates in that
    group.}
 \item{\code{CV}}{: A list of length \code{n_groups}. Each element of
  \code{CV} should be a scalar of vector giving the coefficient o variation
   of the delay(s) to use for simulation of dates in that group.}
 \item{\code{zeta}}{: A scalar in [0;1] giving the probability that, if a
  data point is not missing, it is recorded with error.}
}
\code{hyperparameters} should be a list containing:
\itemize{
 \item{\code{shape1_prob_error}}{: A scalar giving the first shape parameter
  for the beta prior used for parameter \code{theta$zeta}}
 \item{\code{shape2_prob_error}}{: A scalar giving the second shape
  parameter for the beta prior used for parameter \code{theta$zeta}}
 \item{\code{mean_mean_delay}}{: A scalar giving the mean of the exponential
  prior used for parameter \code{theta$mu}}
 \item{\code{mean_CV_delay}}{: A scalar giving the mean of the exponential
  prior used for parameter \code{theta$CV}}
}
\code{index_dates} should be a list of length
 \code{n_groups = length(obs_dat)}. Each element of \code{index_dates}
  should be a matrix with 2 rows and a number of columns corresponding to
   the delays of interest for that group. For each column (i.e. each delay),
    the first row gives the index of the origin date, and the second row
     gives the index of the destination date.
The number of columns of index_dates[[k]] should match the length of
 theta$mu[[k]] and theta$CV[[k]]

If index_dates[[k]] has two columns containing respectively c(1, 2) and
 c(1, 3), this indicates that theta$mu[[k]] and theta$CV[[k]] are
  respectively the mean and coefficient of variation of two delays: the
   first delay being between date 1 and date 2, and the second being between
    date 1 and date 3.

The function performs the move as follows, using a Metropolis-Hastings
 algorithm.
If E=-1 nothing happens.
If E=1, we propose a move to E=0 and hence D=the observed data.
If E=0, we propose a move to E=1. D is then moved as follows: a new value is
 drawn from the marginal posterior of one of the delays this date is
  involved in, repeatedly until D falls on a different day than the observed
   date to be consistent with E=1.
If the date is involved in several delays, one of the delays is randomly
 selected.
This move is not symmetrical so we use a correction factor in computing the
 probability of acceptance in the Metropolis Hastings which accounts for the
  asymetry.
}
\examples{
### TO WRITE OR ALTERNATIVELY REFER TO VIGNETTE TO BE WRITTEN ###
}
