% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{move_lognormal}
\alias{move_lognormal}
\title{Performs one iteration of an MCMC move for either the parameter mu or the parameter CV (mean or CV of the various delays to be estimated)}
\usage{
move_lognormal(what = c("mu", "CV"), group_idx, delay_idx, sdlog, aug_dat,
  curr_theta, obs_dat, hyperpriors, index_dates)
}
\arguments{
\item{what}{A string ("mu" or "CV") indicating which of the parameters to move}

\item{group_idx}{The index of the group for which mu or CV should be moved}

\item{delay_idx}{The index of the delay for which mu or CV should be moved}

\item{sdlog}{The standard deviation to be used for the proposal distribution, see details.}

\item{aug_dat}{The augmented data; a list of observed data, in the format returned by \code{\link{simul_true_data}}.}

\item{curr_theta}{The current list of parameters; see details.}

\item{obs_dat}{A list of observed data, in the format of the first element (called \code{obs_dat}) in the list returned by \code{\link{simul_obs_dat}}.}

\item{hyperpriors}{A list of hyperpriors: see details.}

\item{index_dates}{A list containing indications on which delays to consider in the estimation, see details.}
}
\value{
A list of two elements:
 \itemize{
 \item{\code{new_theta}}{: Same as \code{curr_theta} but where \code{curr_theta$zeta} has been updated}
 \item{\code{accept}}{: A scalar with value 1 if the move was accepted and 0 otherwise}
}
}
\description{
Performs one iteration of an MCMC move for either the parameter mu or the parameter CV (mean or CV of the various delays to be estimated)
}
\details{
\code{curr_theta} should be a list containing:
\itemize{
 \item{\code{mu}}{: A list of length \code{n_groups} (the number of groups to be simulated data). Each element of \code{mu} should be a scalar of vector giving the mean delay(s) to use for simulation of dates in that group.}
 \item{\code{CV}}{: A list of length \code{n_groups}. Each element of \code{CV} should be a scalar of vector giving the coefficient o variation of the delay(s) to use for simulation of dates in that group.}
 \item{\code{zeta}}{: A scalar in [0;1] giving the probability that, if a data point is not missing, it is recorded with error.}
}
\code{hyperpriors} should be a list containing:
\itemize{
 \item{\code{shape1_prob_error}}{: A scalar giving the first shape parameter for the beta prior used for parameter \code{theta$zeta}}
 \item{\code{shape2_prob_error}}{: A scalar giving the second shape parameter for the beta prior used for parameter \code{theta$zeta}}
 \item{\code{mean_mean_delay}}{: A scalar giving the mean of the exponential prior used for parameter \code{theta$mu}}
 \item{\code{mean_CV_delay}}{: A scalar giving the mean of the exponential prior used for parameter \code{theta$CV}}
}

\code{index_dates} should be a list of length \code{n_groups=length(obs_dat)}. Each element of \code{index_dates} should be a matrix with 2 rows and a number of columns corresponding to the delays of interest for that group. For each column (i.e. each delay), the first row gives the index of the origin date, and the second row gives the index of the destination date. 
The number of columns of index_dates[[k]] should match the length of theta$mu[[k]] and theta$CV[[k]] 

If index_dates[[k]] has two columns containing respectively c(1, 2) and c(1, 3), this indicates that theta$mu[[k]] and theta$CV[[k]] are respectively the mean and coefficient of variation of two delays: the first delay being between date 1 and date 2, and the second being between date 1 and date 3. 

The function performs the move as follows, using a Metropolis-Hastings algorithm. 
For the parameter to be moved, a new value is drawn from a lognormal distribution with parameters \code{meanlog} equals the log of the current parameter value, and \code{sdlog=sdlog}.
The new parameter set is then accepted with probability given by the ratio of the posterior values at the new parameter set and the old parameter set, multiplie by a correction factor to reflect the non-symetrical nature of the move.
}
\examples{
### TO WRITE OR ALTERNATIVELY REFER TO VIGNETTE TO BE WRITTEN ###
}

