% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{perform_E1_to_E0_swap}
\alias{perform_E1_to_E0_swap}
\title{Perform the E=1 to E=0 swap for a set of dates}
\usage{
perform_E1_to_E0_swap(
  i,
  group_idx,
  date_idx_E1_to_E0,
  curr_aug_dat,
  theta,
  obs_dat,
  hyperparameters,
  index_dates,
  range_dates
)
}
\arguments{
\item{i}{Index of the individual.}

\item{group_idx}{Index of the group.}

\item{date_idx_E1_to_E0}{Numeric vector of column indices for the dates
that are `E=1` and will be moved.}

\item{curr_aug_dat}{Current augmented data list before the move.}

\item{theta}{List of model parameters (not used in this specific function
but passed for consistency with other swap steps).}

\item{obs_dat}{List of the observed data.}

\item{hyperparameters}{List of model hyperparameters (not used in this
specific function).}

\item{index_dates}{List defining delays for each group (not used in this
specific function).}

\item{range_dates}{Vector of the overall date range (not used in this
specific function).}
}
\value{
The augmented data list (`proposed_aug_dat`) with the relevant dates
  and error indicators updated.
}
\description{
First sequential step within `swap_Ei`. It takes all dates for a given
individual that are currently marked as an error (`E=1`) and proposes a new
state where they are marked as correct (`E=0`).
}
\details{
This is a deterministic proposal. For each date being moved, the error
indicator `E` is set to `0`, and the true date `D` is set to the
corresponding observed date from `obs_dat`. As this move is deterministic,
the forward proposal probability is 1, and the log-probability is 0.
}
\seealso{
`swap_Ei`, `propose_move_from_E1_to_E0`
}
