% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{plot_aug_dat_chains}
\alias{plot_aug_dat_chains}
\title{Plots the MCMC chains of augmented data}
\usage{
plot_aug_dat_chains(MCMCres, aug_dat_true = NULL)
}
\arguments{
\item{MCMCres}{The output of function \code{\link{RunMCMC}}.}

\item{aug_dat_true}{A list containing the data to which the output chains should be compared. If not \code{NULL}, this should have the format of the output of function \code{\link{simul_true_data}}.}
}
\value{
Nothing. Only performs a plot.
}
\description{
Plots the MCMC chains of augmented data
}
\examples{
### TO WRITE OR ALTERNATIVELY REFER TO VIGNETTE TO BE WRITTEN ###
}

