% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulFunctions.R
\name{simul_obs_dat}
\alias{simul_obs_dat}
\title{Introduces missingness and errors in data}
\usage{
simul_obs_dat(D, theta, range_dates, p_error = list(external_swap = 0.04,
  internal_swap = 0.005, neighbour_substitution = 0.05,
  distant_substitution = 0.02, random = 0.01), remove_allNA_indiv = TRUE,
  remove_indiv_at_most_one_date_recorded = TRUE)
}
\arguments{
\item{D}{A list of data, in the format of the first element (called \code{true_dat}) in the list returned by \code{\link{simul_true_data}}.}

\item{theta}{A list of parameters; see details.}

\item{range_dates}{Range of integers in which to draw the erroneous data (these will ve drawn unifromly in that range)}

\item{p_error}{A list with 5 weights defining a multinomial model -see default value. 
The weights inform the probabilities of 'external_swap', 'internal_swap', 'neighbour_substitution' , 'distant_substitution', and 'random' errors. 
If weights don't sum to 1; they are automatically rescaled to define corresponding probabilities.}

\item{remove_allNA_indiv}{A boolean stating whether individuals with only NA dates should be removed or not.}

\item{remove_indiv_at_most_one_date_recorded}{A boolean stating whether individuals with at most one recorded date should be removed or not.}
}
\value{
A list with two elements: 
\itemize{
 \item{\code{obs_dat}}{: A list similar to \code{D}, but where some data points are now missing, and some are erroneous}
 \item{\code{E}}{: A list structured similarly to \code{D} and \code{obs_dat}, containing indicators of where \code{obs_dat} is missing (\code{E=-1}), where \code{obs_dat} is recorded but with error (\code{E=1}), and where \code{obs_dat} is recorded with no error (\code{E=0})}
}
}
\description{
Introduces missingness and errors in data
}
\details{
\code{theta} should be a list containing
\itemize{
 \item{\code{prop_missing_data}}{: A scalar in [0;1] giving the probability of each data point being missing.}
 \item{\code{zeta}}{: A scalar in [0;1] giving the probability that, if a data point is not missing, it is recorded with error.}
}
}
\examples{
### Number of groups of individuals to simulate ###
n_groups <- 2
### Number of dates to simulate for each group ###
n_dates <- c(2, 3)
### Setting up the parameters for the simulation ###
theta <- list()
theta$mu <- list(5, c(10, 15)) # mean delays, for each group
theta$CV <- list(0.5, c(0.5, 0.5)) # coefficient of variation of these delays
theta$prop_missing_data <- 0.25 # probability of data missing in observations
theta$zeta <- 0.05 # probability that, when not missing, the date is recorded with error
### Number of individuals to simulate in each group ###
n_per_group <- rep(10, n_groups)
### Range of dates in which to draw the first set of dates for each group ###
range_dates <- date_to_int(c(as.Date("01/01/2014", "\%d/\%m/\%Y"), as.Date("01/01/2015", "\%d/\%m/\%Y")))
### Which delays to use to simulate subsequent dates from the first, in each group? ###
index_dates <- list(matrix(c(1, 2), nrow=2), cbind(c(1, 2), c(1, 3)))
### Perform the simulation ###
D <- simul_true_data(theta, n_per_group, range_dates, index_dates)
observed_D <- simul_obs_dat(D$true_dat, theta, range_dates, remove_allNA_indiv=TRUE)
### the observed dataset is smaller than the true one 
### (some individuals with only missing data do not appear in the observed dataset)
nrow(D$true_dat[[1]])
nrow(observed_D$obs_dat[[1]])
}
